multi-agent-research/requirements.txt
langgraph 
langchain-openai 
streamlit 
requests 
beautifulsoup4 
python-dotenv